Collapse frequency mapping to utils & 1mo/3mo -> 1m/3m
