+++ src/app/api/analytics/metric/[metric_id]/[frequency]/route.ts
@@ -2,6 +2,7 @@ import prisma from "@/app/lib/prisma";
 import Tenant from "@/lib/tenant/tenant";
 import { NextResponse, type NextRequest } from "next/server";
 import { authenticateApiUser } from "@/app/lib/auth/serverAuth";
+import { frequencyToDateAndSQLcrit } from "@/app/api/common/utils/frequencyHandling";
 import { cache } from "react";
 
 type MetricValue = {
@@ -10,48 +11,6 @@ type MetricValue = {
   ts: number;
   value: number;
 };
-
-// TODO - merge this with the code used in Token Balances, after that
-//        PR is merged.
-function frequencyToDates(
-  frequency: string,
-  timeCol: string
-): {
-  lookback: number;
-  skipCrit: string;
-} {
-  const periodLowerCase = frequency.toLowerCase();
-
-  let lookback: number;
-  let skipCrit: string;
-
-  switch (periodLowerCase) {
-    case "24h":
-      lookback = 90;
-      skipCrit = "1=1";
-      break;
-    case "7d":
-      lookback = 180;
-      skipCrit = `extract(DOW from (${timeCol}) = extract(DOW from current_date)`;
-      break;
-    case "1mo":
-      lookback = 365;
-      skipCrit = `extract(DAY from (${timeCol}) = 1`;
-      break;
-    case "3mo":
-      lookback = 365;
-      skipCrit = `extract(DAY from ${timeCol}) = 1 AND mod(extract(MONTH from ${timeCol}), 3) = 0`;
-      break;
-    case "1y":
-      lookback = 365 * 2;
-      skipCrit = `extract(DAY from ${timeCol}) = 31 AND extract(MONTH from ${timeCol}) = 12`;
-      break;
-    default:
-      throw new Error("Invalid frequency value");
-  }
-
-  return { lookback, skipCrit };
-}
 
 async function getMetricTS(metricId: string, frequency: string) {
   const { namespace } = Tenant.current();
@@ -76,7 +35,10 @@ async function getMetricTS(metricId: string, frequency: string) {
     );
   }
 
-  const { lookback, skipCrit } = frequencyToDates(frequency, "block_date");
+  const { lookback, skipCrit } = frequencyToDateAndSQLcrit(
+    frequency,
+    "block_date"
+  );
 
   const QRY = `SELECT block_date AS day,
                       TO_CHAR(block_date, 'YYYY-MM-DD') date,
